# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IntellectFlow - Multi-Agent Orchestration Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type AgentConfigAgentType =
  (typeof AgentConfigAgentType)[keyof typeof AgentConfigAgentType];

export const AgentConfigAgentType = {
  researcher: "researcher",
  writer: "writer",
  coder: "coder",
  analyst: "analyst",
  designer: "designer",
  qa_tester: "qa_tester",
} as const;

export interface AgentConfig {
  agentType: AgentConfigAgentType;
  model?: string;
  temperature?: number;
  maxTokens?: number;
  systemPrompt?: string;
  tools?: string[];
  enabled: boolean;
}

export type MissionStatus = (typeof MissionStatus)[keyof typeof MissionStatus];

export const MissionStatus = {
  draft: "draft",
  running: "running",
  paused: "paused",
  completed: "completed",
  failed: "failed",
  aborted: "aborted",
} as const;

export type MissionPriority =
  (typeof MissionPriority)[keyof typeof MissionPriority];

export const MissionPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Mission {
  id: number;
  name: string;
  description: string;
  status: MissionStatus;
  category: string;
  priority: MissionPriority;
  progress: number;
  agents: AgentConfig[];
  estimatedCost?: number;
  tokensUsed?: number;
  elapsedSeconds?: number;
  createdAt: string;
  updatedAt: string;
  launchedAt?: string | null;
  completedAt?: string | null;
}

export type CreateMissionInputPriority =
  (typeof CreateMissionInputPriority)[keyof typeof CreateMissionInputPriority];

export const CreateMissionInputPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateMissionInput {
  name: string;
  description: string;
  category: string;
  priority: CreateMissionInputPriority;
  agents: AgentConfig[];
}

export type UpdateMissionInputStatus =
  (typeof UpdateMissionInputStatus)[keyof typeof UpdateMissionInputStatus];

export const UpdateMissionInputStatus = {
  draft: "draft",
  running: "running",
  paused: "paused",
  completed: "completed",
  failed: "failed",
  aborted: "aborted",
} as const;

export interface UpdateMissionInput {
  name?: string;
  description?: string;
  status?: UpdateMissionInputStatus;
  progress?: number;
  agents?: AgentConfig[];
}

export interface ActivityLog {
  id: number;
  missionId: number;
  agentType: string;
  action: string;
  details?: string | null;
  createdAt: string;
}

export interface AnalyticsOverview {
  activeMissions: number;
  completedMissions: number;
  failedMissions: number;
  totalMissions: number;
  agentsRunning: number;
  tasksCompletedToday: number;
  tasksCompletedYesterday: number;
  tokensUsedToday: number;
  tokensLimit: number;
  recentActivity: ActivityLog[];
}

export type ListMissionsParams = {
  status?: ListMissionsStatus;
  limit?: number;
};

export type ListMissionsStatus =
  (typeof ListMissionsStatus)[keyof typeof ListMissionsStatus];

export const ListMissionsStatus = {
  running: "running",
  paused: "paused",
  completed: "completed",
  failed: "failed",
  all: "all",
} as const;

export type GetMissionActivityParams = {
  limit?: number;
};

export type GetRecentActivityParams = {
  limit?: number;
};

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  role: string;
  timezone: string;
  avatarUrl?: string | null;
  notifyMissionComplete: boolean;
  notifyMissionFailed: boolean;
  notifyAgentActivity: boolean;
  notifyWeeklyReport: boolean;
  twoFactorEnabled: boolean;
  defaultModel: string;
  fallbackModel: string;
  maxTokensPerRun: number;
  temperature: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateProfileInput {
  name?: string;
  email?: string;
  role?: string;
  timezone?: string;
  avatarUrl?: string;
  notifyMissionComplete?: boolean;
  notifyMissionFailed?: boolean;
  notifyAgentActivity?: boolean;
  notifyWeeklyReport?: boolean;
  twoFactorEnabled?: boolean;
  defaultModel?: string;
  fallbackModel?: string;
  maxTokensPerRun?: number;
  temperature?: string;
}

export interface ApiKey {
  id: number;
  name: string;
  service: string;
  keyMasked: string;
  isActive: boolean;
  lastUsedAt?: string | null;
  createdAt: string;
}

export interface CreateApiKeyInput {
  name: string;
  service: string;
  key: string;
}

export interface UpdateApiKeyInput {
  name?: string;
  isActive?: boolean;
}