# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IntellectFlow - Multi-Agent Orchestration Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: number;
  missionId: number;
  agentType: string;
  action: string;
  details?: string | null;
  createdAt: string;
}