# lib/api-zod/src/generated/types/agentConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IntellectFlow - Multi-Agent Orchestration Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentConfigAgentType } from "./agentConfigAgentType";

export interface AgentConfig {
  agentType: AgentConfigAgentType;
  model?: string;
  temperature?: number;
  maxTokens?: number;
  systemPrompt?: string;
  tools?: string[];
  enabled: boolean;
}