# lib/api-zod/src/generated/types/analyticsOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IntellectFlow - Multi-Agent Orchestration Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityLog } from "./activityLog";

export interface AnalyticsOverview {
  activeMissions: number;
  completedMissions: number;
  failedMissions: number;
  totalMissions: number;
  agentsRunning: number;
  tasksCompletedToday: number;
  tasksCompletedYesterday: number;
  tokensUsedToday: number;
  tokensLimit: number;
  recentActivity: ActivityLog[];
}