# lib/api-zod/src/generated/types/createMissionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IntellectFlow - Multi-Agent Orchestration Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentConfig } from "./agentConfig";
import type { CreateMissionInputPriority } from "./createMissionInputPriority";

export interface CreateMissionInput {
  name: string;
  description: string;
  category: string;
  priority: CreateMissionInputPriority;
  agents: AgentConfig[];
}