# lib/api-zod/src/generated/types/mission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IntellectFlow - Multi-Agent Orchestration Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentConfig } from "./agentConfig";
import type { MissionPriority } from "./missionPriority";
import type { MissionStatus } from "./missionStatus";

export interface Mission {
  id: number;
  name: string;
  description: string;
  status: MissionStatus;
  category: string;
  priority: MissionPriority;
  progress: number;
  agents: AgentConfig[];
  estimatedCost?: number;
  tokensUsed?: number;
  elapsedSeconds?: number;
  createdAt: string;
  updatedAt: string;
  launchedAt?: string | null;
  completedAt?: string | null;
}