# lib/api-zod/src/generated/types/updateMissionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IntellectFlow - Multi-Agent Orchestration Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentConfig } from "./agentConfig";
import type { UpdateMissionInputStatus } from "./updateMissionInputStatus";

export interface UpdateMissionInput {
  name?: string;
  description?: string;
  status?: UpdateMissionInputStatus;
  progress?: number;
  agents?: AgentConfig[];
}